rewrite the base URL in the app shell to allow offline reload on any path
When the app shell is loaded by the service worker, the base URL is rewritten in the same way as it's done when the app shell is generated on the backend. Regardless of the URL used to load the app shell all relative URLs keep pointing to the same resources.

When the app shell is loaded via `/`, the base URL is `.` => relative links like 'sw.js' resolve to `/sw.js`.
When the app shell is loaded via `/my/path`, the base URL is `./..` => relative links like 'sw.js' still resolve to `/sw.js`.