prototype for @Cacheable endpoints
This prototype shows how adding a `@Cacheable` annotation either on an endpoint class or on individual endpoint methods enables the endpoint calls to be cached by the generated Service Worker. Once that happens these endpoints can be called while offline and the last successful response would be replayed and used by app.

Implementation-wise this is a throw-away prototype.